finals-fixes
