+++ src/API/requestHelpers.ts
@@ -1,10 +1,10 @@
-import { IRequestHelpers, TData } from '../variables/tsVariables';
+import { TTasks } from '../types/typesAndInterfaces';
+import { IRequestHelpers } from '../types/typesAndInterfaces';
 import { instance } from './axiosInstance';
 
 export const getAllTodo = async () => {
   try {
     const { data } = await instance.get<IRequestHelpers>('/tasks');
-
     return data;
   } catch (error) {
     console.log(error);
@@ -13,8 +13,7 @@ export const getAllTodo = async () => {
 
 export const postTodo = async (text: string, status: string) => {
   try {
-    const { data } = await instance.post<IRequestHelpers>('/create-task', { text, status });
-    return data;
+    await instance.post<IRequestHelpers>('/create-task', { text, status });
   } catch (error) {
     console.log(error);
   }
@@ -22,8 +21,7 @@ export const postTodo = async (text: string, status: string) => {
 
 export const deleteTodoById = async (id: string) => {
   try {
-    const { data } = await instance.delete<IRequestHelpers>(`/task/${id}`);
-    return data;
+    await instance.delete<IRequestHelpers>(`/task/${id}`);
   } catch (error) {
     console.log(error);
   }
@@ -31,18 +29,15 @@ export const deleteTodoById = async (id: string) => {
 
 export const deleteAllCompletedTodo = async (arrayOfIds: string) => {
   try {
-    const { data } = await instance.delete<IRequestHelpers>(`/tasks/${arrayOfIds}`);
-
-    return data;
+    await instance.delete<IRequestHelpers>(`/tasks/${arrayOfIds}`);
   } catch (error) {
     console.log(error);
   }
 };
 
-export const updateTodo = async ({ id, text, status }: IRequestHelpers) => {
+export const updateTodo = async ({ id, text, status }: TTasks) => {
   try {
     const { data } = await instance.put<IRequestHelpers>('/task/update', { id, text, status });
-
     return data;
   } catch (error) {
     console.log(error);
@@ -51,9 +46,7 @@ export const updateTodo = async ({ id, text, status }: IRequestHelpers) => {
 
 export const changeStatus = async ({ ids, status }: IRequestHelpers) => {
   try {
-    const { data } = await instance.put<IRequestHelpers>('/task/change-status', { ids, status });
-
-    return data;
+    await instance.put<IRequestHelpers>('/task/change-status', { ids, status });
   } catch (error) {
     console.log(error);
   }
