+++ src/components/Main/Main.tsx
@@ -2,19 +2,19 @@ import React, { useState } from 'react';
 
 import { Link } from '@mui/material';
 
-import MainAddTask from './components/MainAddTask';
-import MainTasks from './components/MainTasks';
+import MainAddTask from './MainAddTask';
+import MainTasks from './MainTasks';
 import MainFilters from '../MainFilters/MainFilters';
-import { TFiltersObject } from '../../variables/tsVariables';
+import { TFiltersObject } from '../../types/typesAndInterfaces';
+import { filtersItems } from '../../variables';
 
 interface IMain {
   filter: TFiltersObject;
-  setFilter: (e: React.SetStateAction<string>) => void;
-
+  setFilter: (e: React.SetStateAction<TFiltersObject>) => void;
 }
 
 const Main: React.FC<IMain> = () => {
-  const [filter, setFilter] = useState<TFiltersObject>({ id: 1, label: 'All', value: 'all' });
+  const [filter, setFilter] = useState<TFiltersObject>(filtersItems[0]);
 
   return (
     <div className="flex flex-col justify-center items-center mt-8 ">
@@ -23,6 +23,7 @@ const Main: React.FC<IMain> = () => {
       <MainFilters filter={filter} setFilter={setFilter} />
 
       <MainTasks filter={filter} />
+
       <div className="my-14  text-xs text-center text-commonGrey">
         <div>Double-click on todo text to edit a task</div>
         <Link href="https://github.com/Yunost-79">Created by Yunost</Link>
