+++ src/components/Main/MainAddTask.tsx
@@ -1,17 +1,13 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 
-import { useForm, SubmitHandler, Controller } from 'react-hook-form';
+import { useForm, Controller } from 'react-hook-form';
 import { useMutation, useQueryClient } from 'react-query';
 
-import { CommonButton } from '../../UI/Common/CommonButton.styled';
-import { CommonInput } from '../../UI/Common/CommonInput.styled';
-import { postTodo } from '../../../API/requestHelpers';
+import { CommonButton } from '../UI/Common/CommonButton.styled';
+import { CommonInput } from '../UI/Common/CommonInput.styled';
+import { postTodo } from '../../API/requestHelpers';
 
-interface IMainAddTask {
-  todo?: string;
-}
-
-const MainAddTask: React.FC<IMainAddTask> = () => {
+const MainAddTask = () => {
   const [todoValue, setTodoValue] = useState<string>('');
   const [errorValue, setErrorValue] = useState<boolean>(false);
 
@@ -27,9 +23,9 @@ const MainAddTask: React.FC<IMainAddTask> = () => {
     control,
     handleSubmit,
     // formState: { errors },
-  } = useForm<IMainAddTask>({});
+  } = useForm({});
 
-  const onSubmitTodo: SubmitHandler<IMainAddTask> = () => {
+  const onSubmitTodo = () => {
     if (!todoValue) {
       setErrorValue(true);
       return;
@@ -40,7 +36,6 @@ const MainAddTask: React.FC<IMainAddTask> = () => {
       setErrorValue(false);
     }
     setTodoValue('');
-
   };
 
   return (
