+++ src/components/MainFilters/MainFilters.tsx
@@ -1,21 +1,21 @@
-import {useMemo} from 'react'
+import { useMemo } from 'react';
 
-import { TFiltersObject, fitersItems } from "../../variables/tsVariables";
+import { TFiltersObject } from '../../types/typesAndInterfaces';
+import { filtersItems } from '../../variables';
 
 interface IMainFilter {
   filter: TFiltersObject;
-  setFilter: (e: React.SetStateAction<string>) => void;
+
+  setFilter: (e: React.SetStateAction<TFiltersObject>) => void;
 }
 
 const MainFilters: React.FC<IMainFilter> = ({ filter, setFilter }) => {
-
-
-  const handleChooseFilter = (item: string) => {
+  const handleChooseFilter = (item: TFiltersObject) => {
     setFilter(item);
   };
 
   const filteredItems = useMemo(() => {
-    return fitersItems.map((item: TFiltersObject) => {
+    return filtersItems.map((item: TFiltersObject) => {
       return {
         ...item,
         active: filter.id === item.id,
@@ -30,7 +30,7 @@ const MainFilters: React.FC<IMainFilter> = ({ filter, setFilter }) => {
         return (
           <div
             className={`text-lg font-medium cursor-pointer ease-linear duration-200 hover:text-goldenYellow ${
-              item.active ? "text-goldenYellow" : "text-white"
+              item.active ? 'text-goldenYellow' : 'text-white'
             }`}
             onClick={item.onClick}
             key={item.id}
