+++ src/variables.ts
@@ -1,0 +1,11 @@
+export const filtersItems = [
+  { id: 1, label: 'All', value: 'all' },
+  { id: 2, label: 'Active', value: 'active' },
+  { id: 3, label: 'Completed', value: 'completed' },
+];
+
+export enum EVariables {
+  all = "all",
+  active = 'active',
+  completed = 'completed',
+}
